# PERF-#5554: Implement `drop_duplicates` via new `duplicated`