@YarShev 2023-01-27
> btw, why it didn't work with high-level .apply()? if the difference is THAT significant maybe we should give a closer look at the .apply() itself and speed up it instead of avoiding its usage

There are many places in the high-level .apply() where we access the metadata, which might be None so we materialize it and get slowdown. Also, when doing apply with a callable we don't know in advance any of the metadata in most cases. With the new implementation of `duplicated` we know all of the metadata for the resultant frame. That said, of course, we can look into `apply` in terms of perf improvements as part of a separate issue.