@anmyachev 2023-01-31
@YarShev Can you show performance with triggering execution? (since you not only make the operation asynchronous, but also change the implementation)