@dchigarev 2023-02-01
> ```python
> time for modin: 13.61 s. # master
> time for modin: 5.23 s. # PR5587
> ```
> 
> Just called `print("modin_res =", modin_res)` right after `modin_res = modin_df.drop_duplicates()` in the example to trigger execution.

Minding that the first operation in ray is usually slower we could even "improve" the result by adding a warm-up:
```python
NRUNS = 1
execute(md_df.sum()) # warmup
md_res = timeit.timeit(lambda: execute(md_df.drop_duplicates()), number=NRUNS)
```
So the actual results for this branch are:
```
without warmup: 4.11s
with warmup: 1.13s.
```

<details><summary>Complete script</summary>

```python
import modin.pandas as pd
import pandas
import numpy as np
import timeit
from asv_bench.benchmarks.utils.common import execute

# def exetuce(df):
#     pass

NROWS = 1_000_000
RAND_LOW = 0
RAND_HIGH = 10_000_000

data = {
    "col0": np.tile(
        np.concatenate([np.random.randint(RAND_LOW, RAND_HIGH, NROWS // 100 - 1), [1]]),
        100,
    ),
    "col1": np.tile(
        np.concatenate([np.random.randint(RAND_LOW, RAND_HIGH, NROWS // 100 - 1), [1]]),
        100,
    ),
}

md_df = pd.DataFrame(data)
pd_df = pandas.DataFrame(data)

# drop_duplicate: 
#   input_shape = (1_000_000, 2)
#   output_shape = (10_000, 2)

NRUNS = 1
execute(md_df.sum()) # warmup
md_res = timeit.timeit(lambda: execute(md_df.drop_duplicates()), number=NRUNS)
pd_res = timeit.timeit(lambda: pd_df.drop_duplicates(), number=NRUNS)

print(f"Cumulative time of {NRUNS} runs for modin: {md_res}s.") # ~10s.
print(f"Cumulative time of {NRUNS} runs for pandas: {pd_res}s.") # ~0.3s
```

</details>